Refactor and comonize Modal event handler for drawio and handsontable
https://youtrack.weseek.co.jp/issue/GW-7963
- Unify useHandsontableModalLauncherForView and useDrawioModalLauncherForView to useModalLauncherForView
- Define useModalLauncherForView opts type
- Added modalType argument to useModalLauncherForView
- Replace markdown conditionally based on modalType
- Set open modal handler conditionally based on modalType
- Define exported type of SaveByModalType
- Call useModalLauncherForView for every modalType in PageContentsUtilities
- Remove drawio-modal-launcher-for-view.ts and handsontable-modal-launcher-for-view.ts
