+++ apps/app/src/components/Page/PageContentsUtilities.tsx
@@ -1,11 +1,10 @@
 import { useTranslation } from 'next-i18next';
 
 import { useUpdateStateAfterSave } from '~/client/services/page-operation';
-import { useDrawioModalLauncherForView } from '~/client/services/side-effects/drawio-modal-launcher-for-view';
-import { useHandsontableModalLauncherForView } from '~/client/services/side-effects/handsontable-modal-launcher-for-view';
 import { toastSuccess, toastError } from '~/client/util/toastr';
 import { useCurrentPageId } from '~/stores/page';
-
+import { useModalLauncherForView } from '~/client/services/side-effects/modal-launcher-for-view';
+import { SaveByModalType } from '~/interfaces/page-operation';
 
 export const PageContentsUtilities = (): null => {
   const { t } = useTranslation();
@@ -13,26 +12,20 @@ export const PageContentsUtilities = (): null => {
   const { data: pageId } = useCurrentPageId();
   const updateStateAfterSave = useUpdateStateAfterSave(pageId);
 
-  useHandsontableModalLauncherForView({
-    onSaveSuccess: () => {
-      toastSuccess(t('toaster.save_succeeded'));
+  const modalTypes: SaveByModalType[] = Object.values(SaveByModalType);
 
-      updateStateAfterSave?.();
-    },
-    onSaveError: (error) => {
-      toastError(error);
-    },
-  });
+  // Iterate over each modalType and call useModalLauncherForView for every modal type
+  modalTypes.map((modalType) => {
+    useModalLauncherForView(modalType, {
+      onSaveSuccess() {
+        toastSuccess(t('toaster.save_succeeded'));
 
-  useDrawioModalLauncherForView({
-    onSaveSuccess: () => {
-      toastSuccess(t('toaster.save_succeeded'));
-
-      updateStateAfterSave?.();
-    },
-    onSaveError: (error) => {
-      toastError(error);
-    },
+        updateStateAfterSave?.();
+      },
+      onSaveError: (error) => {
+        toastError(error);
+      },
+    })
   });
 
   return null;
